split out transform_queries module
