+++ crates/store/re_tf/src/lib.rs
@@ -141,12 +141,14 @@
 mod component_type_info;
 mod transform_aspect;
 mod transform_forest;
+mod transform_queries;
 mod transform_resolution_cache;
 
 pub use transform_forest::{PinholeTreeRoot, TransformForest, TransformFromToError, TransformInfo};
+pub use transform_queries::{query_view_coordinates, query_view_coordinates_at_closest_ancestor};
 pub use transform_resolution_cache::{
     CachedTransformsForTimeline, PoseTransformArchetypeMap, ResolvedPinholeProjection,
-    TransformResolutionCache, query_view_coordinates, query_view_coordinates_at_closest_ancestor,
+    TransformResolutionCache,
 };
 
 // Re-export the transform frame id types from re_types.
